fix(jupyter/utils/addons): rename webpack config to .mts
.ts extension defers to the nearest package.json "type" field for
NodeNext module-format detection, but this package only ships
package.json5 (no literal package.json), so tsc defaulted the file
to CommonJS and flagged the required import.meta.url usage
(TS1470). Renaming to .mts makes the ESM format explicit and
unambiguous to tsc, independent of package.json lookup; webpack-cli
and Node both already support .mts natively. tsc --noEmit and
pnpm test both pass clean.
