+++ jupyter/utils/addons/README.md
@@ -47,7 +47,7 @@ xdg-open dist/index.html   # Linux
 Click **Finish loading** to simulate JupyterLab loading (spinner disappears).
 
 After editing `partial-head.html`, `partial-body.html`, `demo/lab-index.template.html`, or the
-PurgeCSS safelist in `webpack.config.ts`, re-run `pnpm build` and reload the browser tab.
+PurgeCSS safelist in `webpack.config.mts`, re-run `pnpm build` and reload the browser tab.
 
 ### Build Process
 
@@ -68,8 +68,8 @@ CI (`.github/workflows/test-addons.yaml`) runs `pnpm test` when files under `jup
 - `apply.sh`: Script to apply the addons to a JupyterLab during Dockerfile build
 - `partial-head.html`, `partial-body.html`: HTML content injected into JupyterLab `index.html`
 - `build-demo.sh`: Copies the demo template to `dist/`, runs `apply.sh`, then fixes static URLs for local preview
-- `cleanup-webpack-plugin.ts`: Custom webpack plugin for asset cleanup (removes unnecessary files)
-- `webpack.config.ts`: Webpack configuration with enhanced tree-shaking
+- `cleanup-webpack-plugin.mts`: Custom webpack plugin for asset cleanup (removes unnecessary files)
+- `webpack.config.mts`: Webpack configuration with enhanced tree-shaking
 - `dist/pf.css`: Tree-shaken PatternFly CSS file with only the necessary styles (shipped in images)
 - `dist/index.html`: Local demo only, generated by `build-demo.sh` (not tracked or shipped)
 - `test-build.sh`: Script to verify the tree-shaking effectiveness
