+++ jupyter/utils/addons/cleanup-webpack-plugin.mts
